Add files via upload
This assignment was completed as part of my studies in cybersecurity via the 3MTT programme of the federal government of Nigeria. 
It is a security audit of the assets of the company and security measures that they have in place compared to what they should have as best practices in the IT world. 
It also includes recommendations. 
Solomon OLALEYE